a few font cart

+++ backend/src/main/java/com/example/demo/repository/CartRepository.java
@@ -9,5 +9,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface CartRepository extends JpaRepository<Cart, Long> {
-	 Optional<Cart> findByUserId(int userId);
+	 Optional<Cart> findByUserId(Long userId);
 }
